wip - need to sort out regenerate in the credential handler

+++ backends/credhub/src/main/kotlin/org/cloudfoundry/credhub/certificates/DefaultCertificatesHandler.kt
@@ -2,40 +2,51 @@ package org.cloudfoundry.credhub.certificates
 
 import com.google.common.collect.Lists
 import org.cloudfoundry.credhub.ErrorMessages
+import org.cloudfoundry.credhub.PermissionOperation
+import org.cloudfoundry.credhub.PermissionOperation.DELETE
+import org.cloudfoundry.credhub.PermissionOperation.READ
+import org.cloudfoundry.credhub.PermissionOperation.WRITE
 import org.cloudfoundry.credhub.audit.AuditableCredential
 import org.cloudfoundry.credhub.audit.CEFAuditRecord
+import org.cloudfoundry.credhub.auth.UserContextHolder
 import org.cloudfoundry.credhub.credential.CertificateCredentialValue
 import org.cloudfoundry.credhub.domain.CertificateCredentialVersion
 import org.cloudfoundry.credhub.domain.CredentialVersion
 import org.cloudfoundry.credhub.entity.Credential
 import org.cloudfoundry.credhub.exceptions.EntryNotFoundException
+import org.cloudfoundry.credhub.exceptions.ParameterizedValidationException
 import org.cloudfoundry.credhub.generate.GenerationRequestGenerator
 import org.cloudfoundry.credhub.generate.UniversalCredentialGenerator
-import org.cloudfoundry.credhub.permissions.PermissionedCertificateService
 import org.cloudfoundry.credhub.requests.CertificateRegenerateRequest
 import org.cloudfoundry.credhub.requests.CreateVersionRequest
 import org.cloudfoundry.credhub.requests.UpdateTransitionalVersionRequest
+import org.cloudfoundry.credhub.services.DefaultCertificateService
+import org.cloudfoundry.credhub.services.PermissionCheckingService
+import org.cloudfoundry.credhub.utils.CertificateReader
 import org.cloudfoundry.credhub.views.CertificateCredentialView
 import org.cloudfoundry.credhub.views.CertificateCredentialsView
 import org.cloudfoundry.credhub.views.CertificateVersionView
 import org.cloudfoundry.credhub.views.CertificateView
 import org.cloudfoundry.credhub.views.CredentialView
 import org.springframework.stereotype.Service
-import java.util.UUID
+import java.util.*
 
 @Service
 class DefaultCertificatesHandler(
-    private val permissionedCertificateService: PermissionedCertificateService,
-    private val certificateService: CertificateService,
+    private val certificateService: DefaultCertificateService,
     private val credentialGenerator: UniversalCredentialGenerator,
     private val generationRequestGenerator: GenerationRequestGenerator,
-    private val auditRecord: CEFAuditRecord
+    private val auditRecord: CEFAuditRecord,
+    private val permissionCheckingService: PermissionCheckingService,
+    private val userContextHolder: UserContextHolder
 ) : CertificatesHandler {
 
     override fun handleRegenerate(
         credentialUuid: String,
         request: CertificateRegenerateRequest
     ): CredentialView {
+
+        checkPermissionsByUuid(credentialUuid, WRITE)
 
         val existingCredentialVersion = certificateService
             .findByCredentialUuid(credentialUuid)
@@ -46,7 +57,7 @@ class DefaultCertificatesHandler(
             .generate(generateRequest) as CertificateCredentialValue
         credentialValue.isTransitional = request.isTransitional
 
-        val credentialVersion = permissionedCertificateService
+        val credentialVersion = certificateService
             .save(
                 existingCredentialVersion,
                 credentialValue,
@@ -58,20 +69,30 @@ class DefaultCertificatesHandler(
         return CertificateView(credentialVersion)
     }
 
+    // todo: test this
     override fun handleGetAllRequest(): CertificateCredentialsView {
-        val credentialList = permissionedCertificateService.getAll()
+        val credentialList = certificateService.getAll()
+            .filter {
+            permissionCheckingService.hasPermission(
+                userContextHolder.userContext.actor!!,
+                it.name!!,
+                PermissionOperation.READ
+            )
+        }
         val list = convertCertificateCredentialsToCertificateCredentialViews(credentialList)
         auditRecord.addAllCredentials(Lists.newArrayList<AuditableCredential>(credentialList))
         return CertificateCredentialsView(list)
     }
 
     override fun handleGetByNameRequest(name: String): CertificateCredentialsView {
-        val credentialList = permissionedCertificateService.getByName(name)
+        checkPermissionsByName(name, READ)
+        val credentialList = certificateService.getByName(name)
         val list = convertCertificateCredentialsToCertificateCredentialViews(credentialList)
         return CertificateCredentialsView(list)
     }
 
     override fun handleGetAllVersionsRequest(uuidString: String, current: Boolean): List<CertificateView> {
+        checkPermissionsByUuid(uuidString, READ)
         val uuid: UUID
         try {
             uuid = UUID.fromString(uuidString)
@@ -79,13 +100,14 @@ class DefaultCertificatesHandler(
             throw EntryNotFoundException(ErrorMessages.Credential.INVALID_ACCESS)
         }
 
-        val credentialList = permissionedCertificateService.getVersions(uuid, current)
+        val credentialList = certificateService.getVersions(uuid, current)
 
         return credentialList.map { credential -> CertificateView(credential as CertificateCredentialVersion) }
     }
 
     override fun handleDeleteVersionRequest(certificateId: String, versionId: String): CertificateView {
-        val deletedVersion = permissionedCertificateService
+        checkPermissionsByUuid(versionId, DELETE)
+        val deletedVersion = certificateService
             .deleteVersion(UUID.fromString(certificateId), UUID.fromString(versionId))
         return CertificateView(deletedVersion)
     }
@@ -94,6 +116,7 @@ class DefaultCertificatesHandler(
         certificateId: String,
         requestBody: UpdateTransitionalVersionRequest
     ): List<CertificateView> {
+        checkPermissionsByUuid(certificateId, WRITE)
         var versionUUID: UUID? = null
 
         if (requestBody.versionUuid != null) {
@@ -101,7 +124,7 @@ class DefaultCertificatesHandler(
         }
 
         val credentialList: List<CredentialVersion>
-        credentialList = permissionedCertificateService
+        credentialList = certificateService
             .updateTransitionalVersion(UUID.fromString(certificateId), versionUUID)
 
         return credentialList
@@ -109,9 +132,10 @@ class DefaultCertificatesHandler(
     }
 
     override fun handleCreateVersionsRequest(certificateId: String, requestBody: CreateVersionRequest): CertificateView {
+        checkPermissionsByUuid(certificateId, WRITE)
         val certificateCredentialValue = requestBody.value
         certificateCredentialValue.isTransitional = requestBody.isTransitional
-        val credentialVersion = permissionedCertificateService.set(
+        val credentialVersion = certificateService.set(
             UUID.fromString(certificateId),
             certificateCredentialValue
         )
@@ -121,7 +145,7 @@ class DefaultCertificatesHandler(
 
     private fun convertCertificateCredentialsToCertificateCredentialViews(certificateCredentialList: List<Credential>): List<CertificateCredentialView> {
         return certificateCredentialList.map { credential ->
-            val certificateVersions = permissionedCertificateService.getAllValidVersions(credential.uuid!!) as List<CertificateCredentialVersion>
+            val certificateVersions = certificateService.getAllValidVersions(credential.uuid!!) as List<CertificateCredentialVersion>
 
             var signedBy = ""
             if (certificateVersions.isNotEmpty()) {
@@ -133,7 +157,7 @@ class DefaultCertificatesHandler(
             }
 
             val signedCertificates = if (credential.name != null) {
-                permissionedCertificateService.findSignedCertificates(credential.name!!)
+                certificateService.findSignedCertificates(credential.name!!)
             } else {
                 emptyList()
             }
@@ -147,6 +171,26 @@ class DefaultCertificatesHandler(
             }
 
             CertificateCredentialView(credential.name, credential.uuid, certificateVersionViews, signedBy, signedCertificates)
+        }
+    }
+
+    private fun checkPermissionsByName(name: String, permissionOperation: PermissionOperation) {
+        if (!permissionCheckingService.hasPermission(
+                userContextHolder.userContext.actor!!,
+                name,
+                permissionOperation
+            )) {
+            throw EntryNotFoundException(ErrorMessages.Credential.INVALID_ACCESS)
+        }
+    }
+
+    private fun checkPermissionsByUuid(uuid: String, permissionOperation: PermissionOperation) {
+        if (!permissionCheckingService.hasPermission(
+                userContextHolder.userContext.actor!!,
+                UUID.fromString(uuid),
+                permissionOperation
+            )) {
+            throw EntryNotFoundException(ErrorMessages.Credential.INVALID_ACCESS)
         }
     }
 }
